tripletexweb/backend/requirements.txt
cyb-tripletex==0.0.1
Flask==2.1.2
Flask-Cors==3.0.10
gunicorn==20.1.0
requests==2.28.1
python-dotenv==0.20.0
pytest>=6.2.5
google-api-python-client==2.52.0
google-auth-oauthlib==0.5.2
